Add a new firefox_install resource
Convert this recipe / attribute based cookbook to be a resource driven cookbook with backwards compatibility to the old usage.\

Also properly install the 64bit version on 64bit windows releases
